Version 2.2.0
- major refactor of togo plugin (tivo download thread is in new class)
- some logging improvements

With the new ShowInfo class metadata is gathered in 2 places, because a lot of functionality remains in metadata.py. In part it's mostly because I haven't tried to move code I couldn't test and don't fully understand, and in part it's because what is now in ShowInfo is what's needed by the togo plugin.

Note: still only tested so far on linux, so it may or may not run on Windows and the Mac