enhancements to syncerr yaml report file
In TivoCommunity forum discussions @ClearToLand had suggested that it would be nice to see the download rate included in my log, @Dan203 mentioned how I could use VideoReDo to see where the corruption was occurring by using the byte offset into the tivo file (actually the MB offset).
In addition I realized that it would be useful to know which of my tivos the download was from.